Supply distance argument to OCM api calls when appropriate.
When OCM API requests are made from the context of the map view controller we calculate the current map bound region, multiply by 0.75 and convert to kilometers. We supply that value to the OCM API which will limit the charging data returned to only include data relevant for the currently visible map region. Initial tests show drastically improved performance.

References #3